abstracted out reviews routes
